WSHP autosizing WIP

+++ src/ancrec.cpp
@@ -1045,7 +1045,7 @@ int FC basAnc::ancNext( size_t &an, BP * _b)		// (registered) record anchor iter
 // init an to 0.  each call returns FALSE if done or returns an and *_b set to next anchor
 // example:  for (USI an = 0; ancNext( an, &b); )  { ... }
 {
-	while (++an < Nanc)
+	while (static_cast<int>(++an) < Nanc)
 		if (ancs[an])
 		{
 			*_b = ancs[an];     // skip (unexpected) NULL ptr
